Rename for more accurate description.
