+++ tests/routes/test_competences_route.py
